requirements.txt
asgiref==3.9.2
attrs==25.3.0
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
